+++ src/main/java/blusunrize/immersiveengineering/client/gui/BlastFurnaceScreen.java
@@ -8,17 +8,40 @@
 
 package blusunrize.immersiveengineering.client.gui;
 
+import blusunrize.immersiveengineering.api.Lib;
 import blusunrize.immersiveengineering.client.ClientUtils;
+import blusunrize.immersiveengineering.common.blocks.metal.BlastFurnacePreheaterTileEntity;
+import blusunrize.immersiveengineering.common.blocks.stone.BlastFurnaceAdvancedTileEntity;
 import blusunrize.immersiveengineering.common.gui.BlastFurnaceContainer;
 import com.mojang.blaze3d.systems.RenderSystem;
+import net.minecraft.client.resources.I18n;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.util.text.ITextComponent;
 
+import java.util.function.Function;
+
 public class BlastFurnaceScreen extends IEContainerScreen<BlastFurnaceContainer>
 {
+	private static final Function<BlastFurnacePreheaterTileEntity, Boolean> PREHEATER_ACTIVE = tile -> tile.active;
+
 	public BlastFurnaceScreen(BlastFurnaceContainer container, PlayerInventory inventoryPlayer, ITextComponent title)
 	{
 		super(container, inventoryPlayer, title);
+		if(container.tile instanceof BlastFurnaceAdvancedTileEntity)
+			this.xSize = 210;
+	}
+
+	@Override
+	protected void drawGuiContainerForegroundLayer(int mouseX, int mouseY)
+	{
+		if(container.tile instanceof BlastFurnaceAdvancedTileEntity)
+		{
+			String title = I18n.format(Lib.GUI+"blast_furnace.preheaters");
+			int w = this.font.getStringWidth(title)/2;
+			this.font.drawString(title, 175-w, 18, 0xAEAEAE);
+			this.font.drawString(I18n.format(Lib.GUI+"left"), 154, 28, 0xAEAEAE);
+			this.font.drawString(I18n.format(Lib.GUI+"right"), 154, 40, 0xAEAEAE);
+		}
 	}
 
 	@Override
@@ -26,7 +49,17 @@ public class BlastFurnaceScreen extends IEContainerScreen<BlastFurnaceContainer>
 	{
 		RenderSystem.color3f(1.0F, 1.0F, 1.0F);
 		ClientUtils.bindTexture("immersiveengineering:textures/gui/blast_furnace.png");
-		this.blit(guiLeft, guiTop, 0, 0, xSize, ySize);
+		this.blit(guiLeft, guiTop, 0, 0, 176, ySize);
+
+		if(container.tile instanceof BlastFurnaceAdvancedTileEntity)
+		{
+			BlastFurnaceAdvancedTileEntity tile = (BlastFurnaceAdvancedTileEntity)container.tile;
+			this.blit(guiLeft+140, guiTop+11, 176, 32, 70, 46);
+			if(tile.getFromPreheater(true, PREHEATER_ACTIVE, false))
+				this.blit(guiLeft+182, guiTop+27, 200, 22, 10, 10);
+			if(tile.getFromPreheater(false, PREHEATER_ACTIVE, false))
+				this.blit(guiLeft+182, guiTop+39, 200, 22, 10, 10);
+		}
 
 		if(container.state.getLastBurnTime() > 0)
 		{