+++ changelog.md
@@ -1,5 +1,6 @@
 ##### Version 1.15.2-4.1.1-pre
 - Added sawdust flooring! It's a snow-like block made from sawdust! (BluSunrize)
+- Added indicators for attached Preheaters to the Blast Furnace GUI (BluSunrize)
 - Changed recipe for Pressurized Air Tank to use blue dye instead of lapis (BluSunrize)
 - Changed item and fluid routers to not output to unfiltered sides, if a filter exists (BluSunrize)
     - This behavior now matches what is described in the manual